Extract smoke tests out from general specs
