+++ .github/actions/deploy-environment/action.yml
@@ -49,7 +49,7 @@ runs:
         ruby-version: 3.3.1
         bundler-cache: true
 
-    - name: Run RSpec tests
+    - name: Run Smoke Tests
       shell: bash
       run: |
         CLAIMS_EXTERNAL_HOST=$(echo $HOSTNAMES | jq -r '.[0]')
@@ -58,7 +58,6 @@ runs:
         export CLAIMS_EXTERNAL_HOST=$CLAIMS_EXTERNAL_HOST
         export PLACEMENTS_EXTERNAL_HOST=$PLACEMENTS_EXTERNAL_HOST
 
-        bin/bundle exec rspec spec/system/smoke_tests --tag smoke_test -b
+        bin/rspec spec/smoke_tests --tag smoke_test -b
       env:
-        SMOKE_TEST: "true"
         HOSTNAMES: ${{ env.HOSTNAMES }}
