+++ spec/smoke_tests/claims/home_page_spec.rb
@@ -1,7 +1,7 @@
-require "rails_helper"
+require "smoke_tests/smoke_test_helper"
 
 RSpec.describe "Home Page", :smoke_test, service: :claims, type: :system do
-  scenario "User visits the claims homepage" do
+  it "User visits the claims homepage" do
     given_i_am_on_the_start_page
     i_can_see_the_claims_service_name_in_the_header
   end
