+++ spec/smoke_tests/placements/home_page_spec.rb
@@ -1,7 +1,7 @@
-require "rails_helper"
+require "smoke_tests/smoke_test_helper"
 
 RSpec.describe "Home Page", :smoke_test, service: :placements, type: :system do
-  scenario "User visits the placements homepage" do
+  it "User visits the placements homepage" do
     given_i_am_on_the_start_page
     i_can_see_the_placements_service_name_in_the_header
   end
